implement pydantic validation

+++ etl/src/movie_etl/sql/ddl/fact_rating.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS fact_rating (
     user_id INT NOT NULL,
     movie_id INT NOT NULL,
-    rating SMALLINT NOT NULL,
+    rating DECIMAL(2,1) NOT NULL,
     "timestamp" TIMESTAMP NOT NULL,
 
     CONSTRAINT pk_rating PRIMARY KEY (user_id, movie_id),
